# Agent/notes/nuveq-openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/visitors": {
      "get": {
        "operationId": "VisitorsController_findAll",
        "parameters": [
          {
            "name": "after",
            "required": true,
            "in": "query",
            "description": "id",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "first",
            "required": true,
            "in": "query",
            "description": "0-1000",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VisitorArrayResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Maximum can only query 1000 data."
          }
        },
        "tags": [
          "visitors"
        ],
        "security": [
          {
            "X-API-KEY": [
              "X-API-KEY"
            ]
          }
        ]
      },
      "post": {
        "operationId": "VisitorsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateVisitorDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateVisitorResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "visitors"
        ],
        "security": [
          {
            "X-API-KEY": [
              "X-API-KEY"
            ]
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "Nuveq Partner API",
    "description": "",
    "version": "2.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "X-API-KEY": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-KEY"
      }
    },
    "schemas": {
...
      "VisitorDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "name": {
            "type": "string"
          },
          "userId": {
            "type": "string",
            "nullable": true
          },
          "phone": {
            "type": "string",
            "nullable": true
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "userPhoto": {
            "type": "string",
            "nullable": true
          },
          "lastVisit": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          },
          "blacklisted": {
            "type": "boolean",
            "nullable": true
          },
          "remarks": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "id",
          "name"
        ]
      },
      "VisitorArrayResponseDto": {
        "type": "object",
        "properties": {
          "error": {
            "type": "number"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/VisitorDto"
            }
          },
          "cursor": {
            "type": "number"
          },
          "total": {
            "type": "number"
          }
        },
        "required": [
          "error",
          "message",
          "data"
        ]
      },
...
      "CreateVisitorDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Person"
          },
          "credentialNumber": {
            "type": "number",
            "example": 1253646425
          },
          "email": {
            "type": "string",
            "nullable": true,
            "example": "person@email.com"
          },
          "phone": {
            "type": "string",
            "nullable": true,
            "example": "9483014014103"
          },
          "userPhoto": {
            "type": "string",
            "nullable": true,
            "example": "https://examplesite.com/image.png"
          },
          "vehicleNumber": {
            "type": "string",
            "nullable": true,
            "example": "123214"
          },
          "visitStart": {
            "format": "date-time",
            "type": "string",
            "description": "ISO 8601 with timezone format",
            "example": "2025-01-06T08:53:58+08:00"
          },
          "visitEnd": {
            "format": "date-time",
            "type": "string",
            "description": "ISO 8601 with timezone format",
            "example": "2025-01-06T09:53:58+08:00"
          },
          "siteId": {
            "type": "number",
            "example": 10
          },
          "liftGroupId": {
            "type": "number",
            "example": 10
          },
          "allowedDoorIds": {
            "type": "array",
            "items": {
              "type": "number"
            }
          }
        },
        "required": [
          "name",
          "credentialNumber",
          "visitStart",
          "visitEnd",
          "siteId",
          "liftGroupId",
          "allowedDoorIds"
        ]
      },
      "CreateVisitorResult": {
        "type": "object",
        "properties": {
          "visitorId": {
            "type": "number"
          },
          "visitorRegistrationId": {
            "type": "number"
          }
        },
        "required": [
          "visitorId",
          "visitorRegistrationId"
        ]
      },
      "CreateVisitorResponseDto": {
        "type": "object",
        "properties": {
          "error": {
            "type": "number"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/CreateVisitorResult"
          }
        },
        "required": [
          "error",
          "message",
          "data"
        ]
      },
...
    }
  }
}